move positive/negative answers to separate file
