+++ public/index.html
@@ -9,7 +9,6 @@
       name="description"
       content="Determine if you should go out during the corona crisis."
     />
-    <link rel="apple-touch-icon" href="%PUBLIC_URL%/logo192.png" />
 
     <link rel="manifest" href="%PUBLIC_URL%/manifest.json" />
     <link
