+++ src/helper-files/determineConclusion.ts
@@ -1,25 +1,8 @@
 import { Answer, Answers } from "../types/answer";
-
-export const positiveAnswers: string[] = [
-  "yes",
-  "ja",
-  "si",
-  "sim",
-  "yes, by myself or with people I live with",
-  "ja, alleen of alleen met mensen die bij je in huis wonen",
-  "si, solo o con las personas que vivo",
-  "sim, sozinho ou com pessoas com quem vivo"
-];
-
-export const negativeAnswers: string[] = [
-  "no",
-  "nee",
-  "não",
-  "no, with others",
-  "nee, ook met andere mensen",
-  "no, con mas gente",
-  "não, vou sair com outras pessoas"
-];
+import {
+  positiveAnswers,
+  negativeAnswers
+} from "../translations/pos-neg-answers";
 
 export const determineConclusion = (
   finalQuestionAnswer: string,
