+++ src/translations/pos-neg-answers.ts
@@ -1,0 +1,22 @@
+type PositiveNegativeAnswers = string[];
+
+export const positiveAnswers: PositiveNegativeAnswers = [
+  "yes",
+  "ja",
+  "si",
+  "sim",
+  "yes, by myself or with people I live with",
+  "ja, alleen of alleen met mensen die bij je in huis wonen",
+  "si, solo o con las personas que vivo",
+  "sim, sozinho ou com pessoas com quem vivo"
+];
+
+export const negativeAnswers: string[] = [
+  "no",
+  "nee",
+  "não",
+  "no, with others",
+  "nee, ook met andere mensen",
+  "no, con mas gente",
+  "não, vou sair com outras pessoas"
+];
